feat(vite.config): migrate to mjs and seed env variables on the fly
